Bug=44327 author=ashuk r=a=edburns
This Fix enables multiple windows in Webclient on Win32. The fix modifies
the following files:

M mozilla/java/webclient/classes_spec/org/mozilla/webclient/wrapper_native/NativeEventThread.java
M mozilla/java/webclient/classes_spec/org/mozilla/webclient/test/EmbeddedMozilla.java
M mozilla/java/webclient/src_moz/Makefile.win
M mozilla/java/webclient/src_moz/NativeEventThread.cpp
M mozilla/java/webclient/src_moz/NavigationActionEvents.cpp
M mozilla/java/webclient/src_moz/WindowControlActionEvents.cpp
M mozilla/java/webclient/src_moz/WindowControlImpl.cpp
M mozilla/java/webclient/src_moz/ns_util.cpp
M mozilla/java/webclient/src_moz/ns_util.h
? mozilla/java/webclient/src_moz/NativeEventThreadActionEvents.h
? mozilla/java/webclient/src_moz/NativeEventThreadActionEvents.cpp

Pls refer http://bugzilla.mozilla.org/show_bug.cgi?id=44327 for more
info on this Bug and this Fix

+++ mozilla/java/webclient/classes_spec/org/mozilla/webclient/test/EmbeddedMozilla.java
@@ -68,7 +68,7 @@ public void CreateEMWindow()
 {
   System.out.println("Creating new EmbeddedMozilla window");
   EMWindow aEMWindow ;
-  aEMWindow = new EMWindow("EmbeddedMozila#" + count+1,
+  aEMWindow = new EMWindow("EmbeddedMozila#" + (int)(count+1),
                                  binDir, url, count, this);
   count++;
 }
